Update stargazer.dm
Energy output drops slightly from 7 to 5 per second. Hp drops from 40, several hits with any makeshift weapon, to 10, 1-2 hits from makeshift weapons. 

Star gazers take seconds to make and refund their cost very quickly. Spam shouldn't be encouraged, let alone so profitable. Having them be easily broken is appropriate for this kind of spammy, stackable affect, especially one noted in it's description as fragile. The tactic is already to hide row upon row of them someplace out of the way and or difficult to get to. The test bomb satelite, the bottom part of arrivals no one goes to, escape pods, atmosia, perma, and so on. Some of these places will then be vented. Yes, there are ways to deal with them. It just shouldn't be so, literally tiring, and time consuming. Lets keep the art of their use in the positioning, rather than the annoyance of repetitive clean up.